[Rename] 공지사항 상세 페이지 파일명 변경
